requirements.txt
langchain>=0.1.0
langchain-community>=0.0.16
langchain-google-genai>=0.0.5
faiss-cpu>=1.7.4
pypdf>=3.17.0
pandas>=2.0.0
streamlit>=1.32.0
python-dotenv>=1.0.0